requirements.txt
openai
langchain
langchain_chroma
langchain-ollama
embeddings
python-dotenv
sentence-transformers
langchain_community
langchain_text_splitters
python-dotenv
streamlit
pypdf
mistralai
uuid
yt_dlp
pydub
